+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project changes will be shown here.
 ### Added
 - Issue 906 - Sales Order Permissions for Non Sales People
 - Issue 1065 - Include Costs in Contract Renewals Page
+- Issue 1062 - Include email address in Microsoft 365 export
 ### Changed
 - Issue 1025 - Option to not notify Sales on Sales Request Approvals
 ### Fixed